refactor data slice
